requirements.txt
fastapi
uvicorn
gunicorn
PyPDF2
pdf2image
pytesseract
Pillow
opencv-python-headless
pyzbar
numpy
python-multipart